TOT-80 : [REFACTOR] PageResponseDto를 common패키지로 이동
