+++ backend/src/main/java/com/triportreat/backend/plan/service/impl/PlanServiceImpl.java
@@ -4,7 +4,7 @@ import com.triportreat.backend.common.error.exception.AuthenticateFailException;
 import com.triportreat.backend.place.entity.Place;
 import com.triportreat.backend.place.repository.PlaceRepository;
 import com.triportreat.backend.plan.domain.PlanResponseDto.PlanDetailResponseDto;
-import com.triportreat.backend.plan.domain.PageResponseDto;
+import com.triportreat.backend.common.response.PageResponseDto;
 import com.triportreat.backend.plan.domain.PlanRequestDto.PlanCreateRequestDto;
 import com.triportreat.backend.plan.domain.PlanRequestDto.PlanUpdateRequestDto;
 import com.triportreat.backend.plan.domain.PlanRequestDto.ScheduleCreateRequestDto;
