+++ backend/src/test/java/com/triportreat/backend/plan/controller/PlanControllerTest.java
@@ -34,7 +34,7 @@ import com.triportreat.backend.common.config.WebConfig;
 import com.triportreat.backend.common.error.exception.AuthenticateFailException;
 import com.triportreat.backend.dummy.DummyObject;
 import com.triportreat.backend.plan.domain.PlanResponseDto.PlanDetailResponseDto;
-import com.triportreat.backend.plan.domain.PageResponseDto;
+import com.triportreat.backend.common.response.PageResponseDto;
 import com.triportreat.backend.plan.domain.PlanRequestDto.PlanCreateRequestDto;
 import com.triportreat.backend.plan.domain.PlanRequestDto.PlanUpdateRequestDto;
 import com.triportreat.backend.plan.domain.PlanRequestDto.ScheduleCreateRequestDto;
